Fixed several issues and added sample code.
Fixed a bug in which the wrong argument strings were used for
pathnames.

Added sample code to show students how to use the open() system
call and how to report error to standard output.

Removed a comment that would confuse students about how the logic
of the program should work.